feat: unwrap Bus namespace to flat exports + barrel for tree-shaking
Convert `export namespace Bus { ... }` to flat named exports in bus.ts
with an `export * as Bus` barrel in index.ts. This lets bundlers
tree-shake individual Bus exports instead of keeping the entire
namespace as an opaque IIFE.

Also adds the unwrap-namespace automation script and migration spec.
